requirements.txt
Django
numpy
whitenoise
tensorflow
matplotlib
opencv-python
pandas
pillow